refactor: move models stuff to new file
